+++ @angular-generic-table/core/pipes/gt-meta.pipe.ts
@@ -7,7 +7,7 @@ import {GtRowMeta} from '../interfaces/gt-row-meta';
 })
 export class GtMetaPipe implements PipeTransform {
 
-  transform(allRows: Array<GtRow>, page?: number, recordLength?: number, dataLength?: number, metaData?: Array<GtRowMeta>): Array<GtRow> {
+  transform(allRows: Array<GtRow>, page?: number, recordLength?: number): Array<GtRow> {
     for (let i = 0; i < allRows.length; i++) {
       if (!allRows[i].$$gtRowId) {
           const uniqueRowId = '_' + Math.random().toString(36).substr(2, 16);